% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc_mode}
\alias{calc_mode}
\title{Mode of given vector's values}
\usage{
calc_mode(data, return.multiple = TRUE, na.rm = FALSE)
}
\arguments{
\item{data}{vector with given data}

\item{na.rm}{boolean. Should missing data be included or omitted. Default is
FALSE.}

\item{return_multiple}{boolean. If multiple mode's exist, should function
return all of them or only one. Default is TRUE.}
}
\value{
vector with mode.
}
\description{
Calculate mode of given vector's valies.
}
\examples{
#Example of factor column and return.multiple=FALSE
calc_mode(iris$Species, return.multiple=FALSE)
#Example of na.rm = FALSE
calc_mode(c(NA, NA, NA, 2), na.rm = FALSE)
}
