% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate summary statistics}
\usage{
calculate(x, stat, ...)
}
\arguments{
\item{x}{the output from \code{\link{hypothesize}} or \code{\link{generate}}}

\item{stat}{a string giving the type of the statistic to calculate. Current options include "mean", "prop", "diff in means", "diff in props", "chisq", and "F".
or an equation in quotes}

\item{...}{currently ignored}
}
\description{
Calculate summary statistics
}
\examples{

# Permutation test for two binary variables
if (require(dplyr)) {
  mtcars \%>\%
    mutate(am = factor(am), vs = factor(vs)) \%>\%
    specify(am ~ vs) \%>\%
    hypothesize(null = "independence") \%>\%
    generate(reps = 100, type = "permute") \%>\%
    calculate(stat = "diff in props") \%>\%
    visualize()
}
}
