% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randTest.R
\name{conductRandTest}
\alias{conductRandTest}
\alias{randTest}
\alias{randTest,observation,rRandSeq,character-method}
\alias{randTest,observation,rRandSeq,missing-method}
\alias{randTest,observation,randSeq,character-method}
\alias{randTest,observation,randSeq,missing-method}
\title{Conduct a Randomization Test}
\usage{
randTest(obs, refSet, type)

\S4method{randTest}{observation,rRandSeq,character}(obs, refSet, type)

\S4method{randTest}{observation,rRandSeq,missing}(obs, refSet, type)

\S4method{randTest}{observation,randSeq,character}(obs, refSet, type)

\S4method{randTest}{observation,randSeq,missing}(obs, refSet, type)
}
\arguments{
\item{obs}{Observations of the trial containing a randomization sequence and the responses, see \code{\link{genObs}}}

\item{refSet}{Reference set for the present comparison}

\item{type}{character string indicating the type of test statistic for the comparison. Can be 
\describe{
\item{\code{"lrs2"}}{linear rank test statistic,}
\item{\code{"cdr"}}{for centrlized difference in ranks (linear rank test),}
\item{\code{dm}}{for difference in means,}
\item{\code{"se"}}{for sum of group E,}
\item{\code{"sc"}}{for sum of group C,} 
\item{\code{"dmed"}}{for difference in medians, or}
\item{\code{"t"}}{for the t-test statistic.}
}}
}
\value{
Returns the type of test statistic that had been used, the observed value of the test statistic and the the p-value.
}
\description{
Calculates the observed value of the test statistic and compares
it to the values it would have attained if one of the sequences of the
reference set had been observed.
}
\examples{
# from randomizeR
params <- rarPar(10)
rs <- genSeq(params)
endp <- normEndp(c(0,1), c(1,1))
refSet <- genSeq(params, 100)
# new in infeR
obs <- genObs(rs, endp)
randTest(obs, refSet)

}
