% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{conf_int}
\alias{p_value}
\title{Deprecated functions and objects}
\usage{
conf_int(x, level = 0.95, type = "percentile", point_estimate = NULL)

p_value(x, obs_stat, direction)
}
\arguments{
\item{x}{See the non-deprecated function.}

\item{level}{See the non-deprecated function.}

\item{type}{See the non-deprecated function.}

\item{point_estimate}{See the non-deprecated function.}

\item{obs_stat}{See the non-deprecated function.}

\item{direction}{See the non-deprecated function.}
}
\description{
These functions and objects should no longer be used. They will be removed
in a future release of \code{infer}.
}
\seealso{
\code{\link[=get_p_value]{get_p_value()}}, \code{\link[=get_confidence_interval]{get_confidence_interval()}}, \code{\link[=generate]{generate()}}
}
