% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_num.R
\name{describe_num}
\alias{describe_num}
\title{Describe numeric columns}
\usage{
describe_num(df)
}
\arguments{
\item{dataframe}{data.frame}
}
\value{
data.frame of descriptive statistics
}
\description{
Descriptive statistics like missing data ratio, median, 
skewness and others of numeric columns of passed data frame.
}
\examples{
describe_num(iris)
}
