% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.infer}
\alias{fit.infer}
\title{Fit linear models to infer objects}
\usage{
\method{fit}{infer}(object, ...)
}
\arguments{
\item{object}{Output from an infer function---likely \code{\link[=generate]{generate()}} or
\code{\link[=hypothesize]{hypothesize()}}---which specifies the formula and data to fit a model to.}

\item{...}{Any optional arguments to pass along to the chosen computational
engine. See \code{\link[parsnip:linear_reg]{parsnip::linear_reg()}} for more information.}
}
\value{
A \link[tibble:tibble]{tibble} containing the following columns:

\itemize{
\item \code{replicate}: Only supplied if the input object had been previously
passed to \code{\link[=generate]{generate()}}. A number corresponding to which resample of the
original data set the model was fitted to.
\item \code{term}: The explanatory variable (or intercept) in question.
\item \code{estimate}: The model coefficient for the given resample (\code{replicate}) and
explanatory variable (\code{term}).
}
}
\description{
Given the output of an infer core function, this function will fit
a model using \code{\link[stats:glm]{stats::glm()}} according to the formula and data supplied
earlier in the pipeline. If passed the output of \code{\link[=specify]{specify()}} or
\code{\link[=hypothesize]{hypothesize()}}, the function will fit one model. If passed the output
of \code{\link[=generate]{generate()}}, it will fit a model to each data resample, denoted in
the \code{replicate} column.

Learn more in \code{vignette("infer")}.
}
\examples{
# fit a linear model predicting number of hours worked per
# week using respondent age and degree status.
observed_fit <- gss \%>\%
  specify(hours ~ age + college) \%>\%
  hypothesize(null = "independence") \%>\%
  fit()

observed_fit

# fit 100 models to resamples of the gss dataset, where the response 
# `hours` is permuted in each. note that this code is the same as 
# the above except for the addition of the `generate` step.
null_fits <- gss \%>\%
  specify(hours ~ age + college) \%>\%
  hypothesize(null = "independence") \%>\%
  generate(reps = 100, type = "permute") \%>\%
  fit()

null_fits

# more in-depth explanation of how to use the infer package
\dontrun{
vignette("infer")
}  

}
