% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{generateObservation}
\alias{generateObservation}
\alias{genObs}
\alias{genObs,randSeq,endpoint,missing-method}
\alias{genObs,randSeq,numeric,missing-method}
\alias{genObs,randSeq,endpoint,selBias-method}
\alias{genObs,randSeq,endpoint,chronBias-method}
\title{Generates observations of a RCT}
\usage{
genObs(rs, endp, issue)

\S4method{genObs}{randSeq,endpoint,missing}(rs, endp, issue)

\S4method{genObs}{randSeq,numeric,missing}(rs, endp, issue)

\S4method{genObs}{randSeq,endpoint,selBias}(rs, endp, issue)

\S4method{genObs}{randSeq,endpoint,chronBias}(rs, endp, issue)
}
\arguments{
\item{rs}{observed randomization sequence, generated by \code{\link[randomizeR]{genSeq}}}

\item{endp}{\code{\link[randomizeR]{normEndp}} object indicating the endpoint of the responses}

\item{issue}{The observations can be affected by selection bias
(\code{\link[randomizeR]{selBias}}) or chronological bias (\code{\link[randomizeR]{chronBias}})}
}
\value{
An object containing an object of class observation containing a
randomization sequence and the response vector.
}
\description{
Generate observed response corresponding to the observed randomization sequence.
}
\examples{
## generate an observation corresponding to a randomization sequence
rs <- genSeq(rarPar(10))
endp <- normEndp(c(0,1), c(1,1))
# new in infeR
genObs(rs, endp)

## Observations with Selection Bias
issue <- selBias("CS", 4, "exact")
# new in infeR
genObs(rs, endp, issue)

## Observation with chronological Bias with linear time trend
issue <- chronBias("linT", 0.05, "exact")
genObs(rs, endp, issue)


}
