% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{get_ci}
\alias{get_ci}
\alias{conf_int}
\alias{get_confidence_interval}
\title{Compute confidence interval}
\usage{
conf_int(x, level = 0.95, type = "percentile", point_estimate = NULL)

get_ci(x, level = 0.95, type = "percentile", point_estimate = NULL)

get_confidence_interval(x, level = 0.95, type = "percentile",
  point_estimate = NULL)
}
\arguments{
\item{x}{Data frame of calculated statistics or containing attributes of
theoretical distribution values. Currently, dependent on statistics being
stored in \code{stat} column as created in \code{\link[=calculate]{calculate()}} function.}

\item{level}{A numerical value between 0 and 1 giving the confidence level.
Default value is 0.95.}

\item{type}{A string giving which method should be used for creating the
confidence interval. The default is \code{"percentile"} with \code{"se"}
corresponding to (multiplier * standard error) as the other option.}

\item{point_estimate}{A numeric value or a 1x1 data frame set to \code{NULL} by
default. Needed to be provided if \code{type = "se"}.}
}
\value{
A 2 x 1 tibble with values corresponding to lower and upper values in
the confidence interval.
}
\description{
Only simulation-based methods are (currently only) supported.
\code{get_confidence_interval()} and \code{get_ci()} are both aliases of \code{conf_int()}.
}
\examples{
mtcars_df <- mtcars \%>\%
  dplyr::mutate(am = factor(am))
d_hat <- mtcars_df \%>\%
  specify(mpg ~ am) \%>\%
  calculate(stat = "diff in means", order = c("1", "0"))
bootstrap_distn <- mtcars_df \%>\%
  specify(mpg ~ am) \%>\%
  generate(reps = 100) \%>\%
  calculate(stat = "diff in means", order = c("1", "0"))
bootstrap_distn \%>\% conf_int(level = 0.9)
bootstrap_distn \%>\% conf_int(type = "se", point_estimate = d_hat)

}
