% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_types.R
\name{get_data_types}
\alias{get_data_types}
\title{Get data types}
\usage{
get_data_types(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data frame of data type summary
}
\description{
From given data frame get aggregated data frame of all column
data types and count
}
\examples{
df <- data.frame(column1 = rep(TRUE,5),
                 column2 = LETTERS[1:5],
                 column3 = seq(1,5),
                 column4 = c(rep("a",4), "b"),
                 column5 = seq(1,1.4,by=0.1),
                 stringsAsFactors = FALSE)
df$column4 <- as.factor(df$column4)
get_data_types(df)
}
