% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBias.R
\name{getBias}
\alias{getBias}
\title{Calculate a bias of a randomization sequence in convergence strategy.}
\usage{
getBias(R, eta)
}
\arguments{
\item{R}{randomization sequence}

\item{eta}{mean shift due to selection bias}
}
\description{
Calculate a bias of a randomization sequence in convergence strategy.
}
