% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPval.R
\name{getPval}
\alias{getPval}
\title{calculate a exact p-value of non-parametric test}
\usage{
getPval(ind, ts, probs)
}
\arguments{
\item{ind}{index of the sequence of M that was actually used for the trial}

\item{ts}{vector of the test statistics for the sequences in the rows of M}

\item{probs}{vector with the probabilities of the sequences contained in M.}
}
\description{
The p-value is the probability of obaining a test statistic as extreme or 
more extreme than the observed test statistic. More extreme is interpreted as
having a more extreme absolute value than the absolute value of the observed
test statistic.
}
