% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPval.R
\name{getPval_obs_t}
\alias{getPval_obs_t}
\title{calculate a exact p-value for given observed value of the test statistic}
\usage{
getPval_obs_t(t_obs, ts, probs)
}
\arguments{
\item{t_obs}{observed value of the test statistic}

\item{ts}{vector of the test statistics for the sequences in the rows of M}

\item{probs}{vector with the probabilities of the sequences contained in M.}
}
\description{
The p-value is the probability of obaining a test statistic as extreme or 
more extreme than the observed test statistic. More extreme is interpreted as
having a more extreme absolute value than the absolute value of the observed
test statistic.
}
