% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{hypothesis}
\alias{hypothesis}
\title{Specify a null hypothesis}
\usage{
hypothesis(x, null, ...)
}
\arguments{
\item{x}{a data frame that can be coerced into a \code{\link[dplyr]{tbl_df}}}

\item{null}{the null hypothesis}

\item{...}{arguments passed to downstream functions}
}
\description{
Specify a null hypothesis
}
\examples{
if (require(dplyr)) {
  mtcars \%>\%
    select(am) \%>\%
    hypothesis(null = "p = 25") \%>\%
    generate(reps = 100) \%>\%
    calculate(stat = "prop")
}
}
