% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\title{Declare a null hypothesis}
\usage{
hypothesize(x, null, ...)
}
\arguments{
\item{x}{A data frame that can be coerced into a \link[dplyr:tbl_df]{tbl_df}.}

\item{null}{The null hypothesis. Options include "independence" and "point".}

\item{...}{Arguments passed to downstream functions.}
}
\value{
A tibble containing the response (and explanatory, if specified)
variable data with parameter information stored as well.
}
\description{
Declare a null hypothesis
}
\examples{
# Permutation test similar to ANOVA
mtcars \%>\%
  dplyr::mutate(cyl = factor(cyl)) \%>\%
  specify(mpg ~ cyl) \%>\%
  hypothesize(null = "independence") \%>\%
  generate(reps = 100, type = "permute") \%>\%
  calculate(stat = "F")

}
