% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infeR-pckg.R
\docType{package}
\name{infeR-package}
\alias{infeR-package}
\alias{infeR}
\title{Randomization based inference for clinical trials}
\description{
Randomization Based Inference in Clinical Trials
}
\details{
\tabular{ll}{
Package: \tab infeR \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-06-03\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes \cr
}

This package implements functions to conduct randomization tests in clinical trials.
Randomization tests are non-parametric tests that only rely on the randomization distribution
of the test statistic. The randomization distribution is induced by the randomization procedure
of the clinical trial. The infeR package relies heavily on the randomizeR R package. It
implements different test statistics found in the literature and provides a convenient interface
to conduct randomization tests, and simulation studies that use randomization tests.
}
\references{
Uschner, D: Randomization based inference in the presence of selection bias. Submitted to Statistics in Medicine, 2020.
}
\author{
Diane Uschner \email{diane.uschner@gmail.com}
}
