% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nztheileria-package.R
\docType{package}
\name{nztheileria-package}
\alias{nztheileria}
\alias{nztheileria-package}
\title{Performs simulation and inference for vector borne diseases.}
\description{
Performs simulation and inference for vector borne diseases.
}
\details{
\tabular{ll}{
Package: \tab nztheileria\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-01\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

Using the model of Jewell and Brown, JRS Interface, 2015, this package
performs inference on and simulates outbreaks of vector borne diseases
where knowledge of the vector population is sparse.  Specifically, this
model and associated algorithms were designed for analysis of the 2012
outbreak of Theileria orientalis (Ikeda) in New Zealand.
}
\author{
Chris Jewell \email{c.jewell@lancaster.ac.uk}
}
\references{
Jewell CP, Brown RG (2015) Bayesian data assimilation provides rapid decision support for vector borne diseases. \emph{J. Roy. Soc. Interface} In press.
}
\keyword{Bayesian}
\keyword{forecast}
\keyword{mcmc}
\keyword{package}
\keyword{theileria}

