% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\docType{class}
\name{observation-class}
\alias{observation-class}
\title{Observations}
\description{
Class for the the observed outcome of the trial, namely a randomization
sequence and a response.
}
\section{Slots}{

\describe{
\item{\code{rs}}{observed randomization sequence, generated by \code{\link{genSeq}}}

\item{\code{resp}}{observed response}
}}

