% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{Inference functions and the initial data frame}
}
\description{
Like \code{dplyr}, \code{infer} also uses the pipe function, \code{\%>\%}
to turn function composition into a series of imperative statements.
}
