% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randTest.R
\docType{class}
\name{randomizationTest-class}
\alias{randomizationTest-class}
\title{Randomization Tests}
\description{
Randomization tests yield the only valid test for the presence of treatment 
effects.
}
\section{Slots}{

\describe{
\item{\code{pValue}}{Probability of rejection}

\item{\code{obs}}{observation consisting of response and randomization sequence}

\item{\code{type}}{type of test statistic that is used}
}}

\references{
Rosenberger Lachin 2016
}
