% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data with multi-byte separator}
\usage{
read_data(filepath, sep, sbyte_sep = "\t", header = TRUE)
}
\arguments{
\item{filepath}{character string. Desciption to connection object}

\item{sep}{string. Description of multi-byte seperator. Single-byte seperator
also can be used.}

\item{sbyte_sep}{string. Single byte seperator that will be used to replace
multi-byte seperator. Default is \emph{\t}.}

\item{header}{logical. Does data contain header.}
}
\value{
data.frame of read in data set
}
\description{
Read in data set which is seperated with multi-byte symbol. 
This function creates workaround for \emph{data.table::fread} function by 
replacing multi-byte symbol with single-byte symbol and then uses built-in 
functions.
}
\examples{
#MovieTweetings dataset from sidooms github project 
df <- read_data(filepath=paste0("https://raw.githubusercontent.com/sidooms/",
                                "MovieTweetings/master/latest/movies.dat"),
          sep="::", sbyte_sep = "~", header=FALSE) 
}
