% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readObs.R
\name{readObs}
\alias{readObs}
\title{Read entries from a csv file and create an observation object. The file must
contain just two columns, where the first consists of the randomization sequence
and the second of the corresponding responses.}
\usage{
readObs()
}
\description{
Read entries from a csv file and create an observation object. The file must
contain just two columns, where the first consists of the randomization sequence
and the second of the corresponding responses.
}
