% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{To determine which theoretical distribution to fit (if any)}
\usage{
set_params(x)
}
\arguments{
\item{x}{A data frame that can be coerced into a \link[tibble:tibble]{tibble}.}
}
\description{
To determine which theoretical distribution to fit (if any)
}
