% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRT.R
\name{simulateRandTest}
\alias{simulateRandTest}
\alias{simRT}
\alias{simRT,randPar,missing-method}
\alias{simRT,randPar,issue-method}
\title{Simulate Randomization Test}
\usage{
simRT(
  randPar,
  issue,
  endp = normEndp(c(0, 1), c(1, 1)),
  L = 500,
  r = 100,
  alpha = 0.05,
  type = "lrs",
  cores = 1
)

\S4method{simRT}{randPar,missing}(
  randPar,
  issue,
  endp = normEndp(c(0, 1), c(1, 1)),
  L = 500,
  r = 100,
  alpha = 0.05,
  type = "lrs",
  cores = 1
)

\S4method{simRT}{randPar,issue}(
  randPar,
  issue,
  endp = normEndp(c(0, 1), c(1, 1)),
  L = 500,
  r = 100,
  alpha = 0.05,
  type = "lrs",
  cores = 1
)
}
\arguments{
\item{randPar}{a \code{\link[randomizeR]{randPar}} object indicating the randomization 
procedure that should be simulated}

\item{issue}{The observations can be affected by selection bias 
(\code{\link[randomizeR]{issue}}) or missing values}

\item{endp}{\code{\link[randomizeR]{normEndp}} object indicating the endpoint of the responses}

\item{L}{integer that denotes the length of the reference set}

\item{r}{integer that denotes how many repetitions should be conducted}

\item{alpha}{significance level of the tests}

\item{type}{Type of test statistic used}

\item{cores}{Number of threads created to make the computation. Default is 1 - no parallelisation}
}
\value{
proportion of tests that yielded a p-value below the significance level
}
\description{
Simulates repeated randomization tests and calculates
the proportion of tests that yielded a p-value below the significance level.
}
\examples{
randPar <- rarPar(8)
issue <- selBias("CS", 0.05, "sim", 0.05)
simRT(randPar = randPar, issue = issue)

#For different iteration steps with more/less sequences, specify r and L
L <- 100
r <- 200
simRT(randPar = randPar, issue = issue, L = L, r = r)

}
\seealso{
Randomization procedures \link[randomizeR]{randPar}.

Endpoint of the responses \link[randomizeR]{normEndp}.

Responses might be influenced by bias, see \link[randomizeR]{issue}.
}
