% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRT.R
\docType{class}
\name{simulationRT-class}
\alias{simulationRT-class}
\title{Randomization Tests}
\description{
Randomization tests yield the only valid test for the presence of treatment 
effects.
}
\section{Slots}{

\describe{
\item{\code{res}}{Proportion of tests that had a result less than the alpha level.}

\item{\code{randPar}}{Representation of the randomization procedure that is being simulated.}

\item{\code{issue}}{Issue for the comparison. Can be of class selBias, mVal, or \code{"missing"}.}

\item{\code{endp}}{Endpoint of the responses that are generated}

\item{\code{L}}{Size of the MC reference set that is generated for each tet}

\item{\code{r}}{Number of tests that are conducted}

\item{\code{alpha}}{alpha level of the test}

\item{\code{type}}{type of test statistic that is used}
}}

\references{
Rosenberger Lachin 2016
}
