% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_clmn_by_rgx_group.R
\name{split_clmn_by_rgx_group}
\alias{split_clmn_by_rgx_group}
\title{Split column by captured regex groups}
\usage{
split_clmn_by_rgx_group(column = column, regex = regex)
}
\arguments{
\item{column}{vector; column from some dataframe that needs to be 
split by groups from regular expression.}

\item{regex}{character; string containing regular expression.}
}
\value{
data frame with columns as many as groups passed in regex.
}
\description{
Split given column by captured regular expression groups. 
Function returns data frame with as many columns as groups in passed regular 
expression.
}
\examples{
df <- data.frame(column = c("a:1, b:1, c:1", "a:1, b:2, c:1", 
                            "a:1, b:3, c:1", "a:1, b:4, c:1"),
                            stringsAsFactors = FALSE)
regex <- ".*b:(.*?), .*"
split_clmn_by_rgx_group(column = df$column, regex = regex)
}
