% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split data by target's distrubtion}
\usage{
get_data_types(df = df, target_column = "some_column")
}
\arguments{
\item{df}{data.frame}

\item{target_column}{string; name of the column which is assumed as target
column. \code{df} must contain this column.}

\item{ratio}{numeric vector; represents proportions how to split data. 
\code{length(ratio) == n}.}

\item{replace}{boolean; sampling with or without replacement.}
}
\value{
list of \emph{n} data frames with equal target distribtuion
}
\description{
Split data set into \emph{n} sets based on target's distribution
with or without observation repetition. Each set should contain at least one
observation per class to keep needed distribution.
}
\examples{
df <- data.frame(column1 = rep(TRUE,50),
                 column2 = c(LETTERS[1:25], LETTERS[26:2]),
                 column3 = seq(1,50),
                 column4 = c(rep("a",45), rep("b",5)),
                 column5 = seq(1,50,by=1),
                 target_col = c(rep("A",25), rep("B", 25)),
                 stringsAsFactors = FALSE)
split_data(df=df, target_column = "target_col", ratio = c(0.1, 0.4, 0.3, 0.2))
}
