% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data_validation}
\alias{split_data_validation}
\title{Validate data frame before splitting}
\usage{
split_data_validation(df = df, target_column = "target_col", n=10)
}
\arguments{
\item{df}{data.frame}

\item{target_column}{string; name of the column which is assumed as target
column. \code{df} must contain this column.}

\item{n}{integer specifying how many sets should be created.}
}
\description{
Validate data frame before splitting
}
\seealso{
\code{\link{split_data}} for splitting data set into \emph{n} parts.
}
