% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{t_test}
\alias{t_test}
\title{t_test}
\usage{
t_test(data, formula, alternative = "two_sided", ...)
}
\arguments{
\item{data}{a data frame that can be coerced into a \code{\link[tibble]{tibble}}}

\item{formula}{a formula with the response variable on the left and the explanatory on the right}

\item{alternative}{character string specifying the direction of the alternative hypothesis. Options are 
"\code{two_sided}" (default), "\code{greater}", or "\code{less}".}

\item{...}{currently ignored}
}
\description{
t_test
}
