% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStat.R
\name{testStat}
\alias{testStat}
\title{calculate a test statistc}
\usage{
testStat(R, y, type = "lrs")
}
\arguments{
\item{R}{randomisation sequence.}

\item{y}{observed response.}

\item{type}{character string indicating the type of test statistic for the comparison. Can be
\describe{
\item{\code{"cdr"}}{for centralized difference in ranks (linear rank test),}
\item{\code{"lrs"}}{for linear rank test when the responses are already ranked,}
\item{\code{"lrs2"}}{for linear rank test when the responses are not yet ranked,}
\item{\code{"ds"}}{for difference in sums,}
\item{\code{dm}}{for difference in means,}
\item{\code{"se"}}{for sum of group E,}
\item{\code{"sc"}}{for sum of group C,}
\item{\code{"dmed"}}{for difference in medians, or}
\item{\code{"t"}}{for the t-test statistic.}
}}
}
\description{
For a given randomisation sequence and response, calculate a certain test statistic.
For now, we only look at the normalized difference of ranks
}
