% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate summary statistics}
\usage{
calculate(x, stat, ...)
}
\arguments{
\item{x}{the output from \code{\link{hypothesize}} or \code{\link{generate}}}

\item{stat}{a string giving the type of the statistic to create, i.e. "diff in means", "diff in props", etc.
or an equation in quotes}

\item{...}{currently ignored}
}
\description{
Calculate summary statistics
}
\examples{

# Permutation test for two binary variables
if (require(dplyr)) {
  diffs <- mtcars \%>\%
    mutate(am = factor(am), vs = factor(vs)) \%>\%
    select(am, vs) \%>\% 
    hypothesize(null = "independence") \%>\% 
    generate(reps = 100, type = "permute") \%>\%
    calculate(stat = "diff in props")
  test_stat <- mtcars \%>\%
    group_by(vs) \%>\%
    summarize(N = n(), manuals = sum(am)) \%>\%
    mutate(prop = manuals / N) \%>\%
    summarize(diff_prop = diff(prop))
  if (require(ggplot2)) {
    ggplot(data = diffs, aes(x = diffprop)) +
      geom_density() + 
      geom_vline(xintercept = 0, linetype = 3) + 
      geom_vline(data = test_stat, aes(xintercept = diff_prop), color = "red")
  }
}
}
