% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate summary statistics}
\usage{
calculate(x, stat, success = NULL, ...)
}
\arguments{
\item{x}{the output from \code{\link{hypothesize}} or \code{\link{generate}}}

\item{stat}{a string giving the type of the statistic to calculate. Current 
options include "mean", "median", "sd", "prop", "diff in means", 
"diff in medians", "diff in props", "Chisq", "F", and "slope".}

\item{success}{a string specifying which level of the response variable should be
used as a success in bootstrapping a one proportion confidence interval}

\item{...}{to pass options like \code{na.rm = TRUE} into functions like mean, sd, etc.}
}
\description{
Calculate summary statistics
}
\examples{
# Permutation test for two binary variables
if (require(dplyr)) {
  mtcars \%>\%
    mutate(am = factor(am), vs = factor(vs)) \%>\%
    specify(am ~ vs) \%>\%
    hypothesize(null = "independence") \%>\%
    generate(reps = 100, type = "permute") \%>\%
    calculate(stat = "diff in props") \%>\%
    visualize()
}

# Confidence interval using bootstrapping for one proportion
mtcars \%>\%
    specify(response = am) \%>\%
    generate(reps = 100, type = "bootstrap") \%>\%
    calculate(stat = "prop", success = "1")
}
