% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{calculate}
\alias{calculate}
\title{Calculate summary statistics}
\usage{
calculate(x, stat, ...)
}
\arguments{
\item{x}{the output from \code{\link{hypothesize}} or \code{\link{generate}}}

\item{stat}{a string giving the type of the statistic to create, i.e. "diff in means", "diff in props", etc.
or an equation in quotes}

\item{...}{currently ignored}
}
\description{
Calculate summary statistics
}
