% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{chisq_stat}
\alias{chisq_stat}
\title{Tidy chi-squared test statistic}
\usage{
chisq_stat(data, formula, ...)
}
\arguments{
\item{data}{A data frame that can be coerced into a \link[tibble:tibble]{tibble}.}

\item{formula}{A formula with the response variable on the left and the
explanatory on the right.}

\item{...}{Additional arguments for \link[stats:chisq.test]{stats::chisq.test}.}
}
\description{
A shortcut wrapper function to get the observed test statistic for a chisq
test. Uses \link[stats:chisq.test]{stats::chisq.test}, which applies a continuity correction.
}
