% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{chisq_test}
\alias{chisq_test}
\title{Tidy chi-squared test}
\usage{
chisq_test(x, formula, response = NULL, explanatory = NULL, ...)
}
\arguments{
\item{x}{A data frame that can be coerced into a \link[tibble:tibble]{tibble}.}

\item{formula}{A formula with the response variable on the left and the
explanatory on the right.}

\item{response}{The variable name in \code{x} that will serve as the response.
This is alternative to using the \code{formula} argument.}

\item{explanatory}{The variable name in \code{x} that will serve as the
explanatory variable.}

\item{...}{Additional arguments for \link[stats:chisq.test]{chisq.test()}.}
}
\description{
A tidier version of \link[stats:chisq.test]{chisq.test()} for goodness of fit
tests and tests of independence.
}
\examples{
# chi-squared test of independence for college completion
# status depending on one's self-identified income class
chisq_test(gss, college ~ finrela)

# chi-squared goodness of fit test on whether self-identified
# income class follows a uniform distribution
chisq_test(gss,
           response = finrela,
           p = c("far below average" = 1/6,
                 "below average" = 1/6,
                 "average" = 1/6,
                 "above average" = 1/6,
                 "far above average" = 1/6,
                 "DK" = 1/6))

}
\seealso{
Other wrapper functions: 
\code{\link{chisq_stat}()},
\code{\link{observe}()},
\code{\link{prop_test}()},
\code{\link{t_stat}()},
\code{\link{t_test}()}
}
\concept{wrapper functions}
