% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_char.R
\name{describe_char}
\alias{describe_char}
\title{Describe nominal columns}
\usage{
describe_char(df)
}
\arguments{
\item{dataframe}{data.frame}
}
\value{
data.frame of descriptive statistics
}
\description{
Descriptive statistics like missing value ratio and mode for 
character and factor columns. s
}
\examples{
describe_char(iris)
}
