% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_num.R
\name{describe_num}
\alias{describe_num}
\title{Describe numeric columns}
\usage{
describe_num(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame of descriptive statistics
}
\description{
Descriptive statistics like missing data ratio, median, 
skewness and others of numeric (\emph{integer}, \emph{double}, \emph{numeric}
) columns of input data frame. All character, factor and logical columns
are dropped; for these column descriptive statistics function 
\code{describe_char()} has been developed.
}
\examples{
describe_num(iris)
}
