% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.infer}
\alias{fit.infer}
\title{Fit linear models to infer objects}
\usage{
\method{fit}{infer}(object, ...)
}
\arguments{
\item{...}{Currently ignored.}

\item{x}{Output from an infer function---likely \code{\link[=generate]{generate()}} or
\code{\link[=hypothesize]{hypothesize()}}---which specifies the formula and data to fit a model to.}
}
\value{
A tibble. If the input had not been passed to \code{\link[=generate]{generate()}}, the tibble
will contain \code{term} and \code{stat} columns. If it had been, there will also
be a \code{replicate} column.

\itemize{
\item \code{replicate}: Only supplied if the input object had been previously
passed to \code{\link[=generate]{generate()}}. A number corresponding to which resample of the
original data set the model was fitted to.
\item \code{term}: The explanatory variable (or intercept) in question.
\item \code{estimate}: The model coefficient for the given resample (\code{replicate}) and
explanatory variable (\code{term}).
\item \code{stat}: The value of a t-statistic under the null hypothesis that
the regression estimate is non-zero.
}
}
\description{
Given the output of an infer core function, this function will fit
a model according to the formula and data supplied earlier in the
pipeline. If passed the output of \code{\link[=specify]{specify()}} or \code{\link[=hypothesize]{hypothesize()}}, the
function will fit one model. If passed the output of \code{\link[=generate]{generate()}},
it will fit a model to each data resample, denoted in the \code{replicate} column.

Learn more in \code{vignette("infer")}.
}
\examples{

# More in-depth explanation of how to use the infer package
\dontrun{
vignette("infer")
}

}
