% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate resamples, permutations, or simulations}
\usage{
generate(x, reps = 1, type = NULL, ...)
}
\arguments{
\item{x}{A data frame that can be coerced into a \link[tibble:tibble]{tibble}.}

\item{reps}{The number of resamples to generate.}

\item{type}{Currently either \code{bootstrap}, \code{permute}, or \code{simulate}
(see below)}

\item{...}{Currently ignored.}
}
\value{
A tibble containing \code{rep} generated datasets, indicated by the
\code{replicate} column.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

Generation creates a null distribution from \code{\link[=specify]{specify()}} and (if needed)
\code{\link[=hypothesize]{hypothesize()}} inputs
}
\section{Generation Types}{


The \code{type} argument determines the method used to create the null
distribution.

\itemize{
\item \code{bootstrap}: A bootstrap sample will be drawn for each \code{rep},
where a sample of size equal to the input sample size is drawn (with
replacement) from the input sample data.
\item \code{permute}: For each \code{rep}, each input value will be randomly
reassigned (without replacement) to a new output value in the sample.
\item \code{simulate}: A value will be sampled from a theoretical distribution
with parameters specified in \code{\link[=hypothesize]{hypothesize()}} for each \code{rep}. (This option is
currently only applicable for testing point estimates.)
}
}

\examples{
# Permutation test for two binary variables
mtcars \%>\%
  dplyr::mutate(am = factor(am), vs = factor(vs)) \%>\%
  specify(am ~ vs, success = "1") \%>\%
  hypothesize(null = "independence") \%>\%
  generate(reps = 100, type = "permute")

}
