% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{hypothesize}
\alias{hypothesize}
\title{Specify a null hypothesis}
\usage{
hypothesize(x, null, ...)
}
\arguments{
\item{x}{a data frame that can be coerced into a \code{\link[dplyr]{tbl_df}}}

\item{null}{the null hypothesis}

\item{...}{arguments passed to downstream functions}
}
\description{
Specify a null hypothesis
}
\examples{
if (require(dplyr)) {

# One binary variable
  mtcars \%>\%
    select(am) \%>\%
    hypothesize(null = "p = 25") \%>\%
    generate(reps = 100) \%>\%
    calculate(stat = "prop")
    
# Permutation test
  mtcars \%>\%
    select(mpg, cyl) \%>\%
    hypothesize(null = "rho = 0") \%>\%
    generate(reps = 100, type = "permute") \%>\%
    calculate(stat = "cor")
}

# Compare with
if (require(dplyr) && require(broom)) {
  cars <- mtcars \%>\%
    summarize(N = n(), num_manual = sum(am))
  with(cars, prop.test(num_manual, N, correct = FALSE)) \%>\%
    tidy()  
}
}
