% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infeR-pckg.R
\docType{package}
\name{infeR-package}
\alias{infeR-package}
\alias{infeR}
\title{Randomization based inference for clinical trials}
\description{
Randomization based inference for clinical trials
}
\details{
\tabular{ll}{
Package: \tab infeR \cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-06-03\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes \cr
}

This package provides functionality for randomization based inference in clinical trials
}
\references{
Conditional randomization tests in clinical trials
}
\author{
Diane Uschner \email{diane.uschner@gmail.com}
}
