% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\docType{package}
\name{infer}
\alias{infer}
\alias{infer-package}
\title{infer: a grammar for statistical inference}
\description{
The objective of this package is to perform statistical inference using a
grammar that illustrates the underlying concepts and a format that coheres
with the tidyverse.
}
\examples{
# Example usage:
library(infer)

}
