\name{nztheileria-package}
\alias{nztheileria-package}
\alias{nztheileria}
\docType{package}
\title{
NZ Theileria inference and simulation package.
}
\description{
This package contains an MCMC algorithm and Gillespie simulation for the
vector-borne disease model of Jewell CP, Brown RG (2015) Bayesian data
assimilation provides rapid decision support for vector borne
diseases. J. Roy. Soc. Interface. In Press.
}
\details{
\tabular{ll}{
Package: \tab nztheileria\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-29\cr
License: \tab GNU Public License version 3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Chris Jewell (c) 2015

Maintainer: Chris Jewell <c.jewell@lancaster.ac.uk>
}
\references{
Jewell CP, Brown RG (2015) Bayesian data assimilation provides rapid
decision support for vector borne diseases. J. Roy. Soc. Interface. In Press.
}
\keyword{ package }
\seealso{
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
