% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{Inference functions and the initial data frame.}
}
\description{
Like \{dplyr\}, \{infer\} also uses the pipe (\code{\%>\%}) function
from \code{magrittr} to turn function composition into a series of
iterative statements.
}
