% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.infer}
\alias{print.infer}
\title{Print methods}
\usage{
\method{print}{infer}(x, ...)
}
\arguments{
\item{x}{An object of class \code{infer}, i.e. output from \code{\link[=specify]{specify()}} or
\code{\link[=hypothesize]{hypothesize()}}.}

\item{...}{Arguments passed to methods.}
}
\description{
Print methods
}
