% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.infer}
\alias{print.infer}
\alias{print.infer_layer}
\alias{print.infer_dist}
\title{Print methods}
\usage{
\method{print}{infer}(x, ...)

\method{print}{infer_layer}(x, ...)

\method{print}{infer_dist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{infer}, i.e. output from \code{\link[=specify]{specify()}} or
\code{\link[=hypothesize]{hypothesize()}}, or of class \code{infer_layer}, i.e. output from
\code{\link[=shade_p_value]{shade_p_value()}} or \code{\link[=shade_confidence_interval]{shade_confidence_interval()}}.}

\item{...}{Arguments passed to methods.}
}
\description{
Print methods
}
