% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.infer}
\alias{print.infer}
\title{Print methods}
\usage{
\method{print}{infer}(x, ...)
}
\arguments{
\item{x}{an object of class \code{infer}, i.e. output from \link{specify}
or \link{hypothesize}}

\item{...}{arguments passed to methods}
}
\description{
Print methods
}
