% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/visualize.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{ggplot_add}
\title{Objects exported from other packages}
\details{
Read more about infer's \link[=fit.infer]{here} function \link[=fit.infer]{here} or
by running \code{?fit.infer} in your console.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggplot_add}}}
}}

