% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_sample_n.R
\name{rep_sample_n}
\alias{rep_sample_n}
\title{Perform repeated sampling}
\usage{
rep_sample_n(tbl, size, replace = FALSE, reps = 1, prob = NULL)
}
\arguments{
\item{tbl}{data frame of population from which to sample}

\item{size}{sample size of each sample}

\item{replace}{should sampling be with replacement?}

\item{reps}{number of samples of size n = \code{size} to take}

\item{prob}{a vector of probability weights for obtaining the elements of 
the vector being sampled.}
}
\value{
A tibble of size \code{rep} times \code{size} rows corresponding to 
\code{rep} samples of size n = \code{size} from \code{tbl}.
}
\description{
Perform repeated sampling of samples of size n. Useful for creating
sampling distributions
}
\examples{
library(dplyr)
library(ggplot2)

# Create a virtual population of N = 2400 balls, of which 900 are red and the
# rest are white
N <- 2400
population <- data_frame(
  ball_ID = 1:N,
  color = c(rep("red", 900), rep("white", N - 900))
)
population

# Take samples of size n = 50 balls; do this 1000 times
samples <- population \%>\%
  rep_sample_n(size = 50, reps = 1000)
samples

# Compute p_hats for all 1000 samples = proportion red
p_hats <- samples \%>\% 
  group_by(replicate) \%>\% 
  summarize(prop_red = mean(color == "red"))
p_hats

# Plot sampling distribution
ggplot(p_hats, aes(x = prop_red)) + 
  geom_histogram(binwidth = 0.05) + 
  labs(x = "p_hat", y = "Number of samples", 
  title = "Sampling distribution of p_hat based on 1000 samples of size n = 50")
}
