% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify}
\alias{specify}
\title{Specify the response and explanatory variables}
\usage{
specify(x, formula, response = NULL, explanatory = NULL)
}
\arguments{
\item{x}{a data frame that can be coerced into a \code{\link[tibble]{tbl_df}}}

\item{formula}{a formula with the response variable on the left and the explanatory on the right}

\item{response}{the variable name in \code{x} that will serve as the response. This is alternative to using the \code{formula} argument.}

\item{explanatory}{the variable name in \code{x} that will serve as the explanatory variable}
}
\description{
Specify the response and explanatory variables
}
