% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_set.R
\name{split_data_validation}
\alias{split_data_validation}
\title{Validate data frame before splitting}
\usage{
split_data_validation(df = df, target_column = "target_col", n=10)
}
\arguments{
\item{df}{data.frame}

\item{target_column}{string; name of the column which is assumed as target
column. \code{df} must contain this column.}

\item{n}{integer;}
}
\description{

}
\examples{
df <- data.frame(column1 = rep(TRUE,50),
                 column2 = LETTERS[1:50],
                 column3 = seq(1,50),
                 column4 = c(rep("a",45), rep("b",5)),
                 column5 = seq(1,50,by=1),
                 target_col = c(rep("A",25), rep("B", 25)),
                 stringsAsFactors = FALSE)
split_data_validation(df, target_column = "target_col", n=10)
}
