% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{t_stat}
\alias{t_stat}
\title{Tidy t-test statistic}
\usage{
t_stat(data, formula, ...)
}
\arguments{
\item{data}{a data frame that can be coerced into a \link[tibble:tibble]{tibble}}

\item{formula}{a formula with the response variable on the left and the
explanatory on the right}

\item{...}{pass in arguments to {infer} functions}
}
\description{
A shortcut wrapper function to get the observed test statistic for a t test.
}
