% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{(Currently) Visualize the randomization-based distribution 
(To be updated to include theory-based distributions)}
\usage{
visualize(df, bins = 30, ...)
}
\arguments{
\item{df}{the output from \code{\link{calculate}}}

\item{bins}{the number of bins in the histogram}

\item{...}{currently ignored}
}
\value{
A ggplot object showing the randomization-based distribution as a histogram.
Preferable to use the ggplot2 package directly here as a histogram does not always
display the distribution well.
}
\description{
(Currently) Visualize the randomization-based distribution 
(To be updated to include theory-based distributions)
}
