% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{(Currently) Visualize the resampling distribution (To be updated to include theory-based distributions)}
\usage{
visualize(df, ...)
}
\arguments{
\item{df}{the output from \code{\link{calculate}}}

\item{...}{currently ignored}
}
\description{
(Currently) Visualize the resampling distribution (To be updated to include theory-based distributions)
}
\examples{

}
